pyproject.toml
[tool.poetry]
name = "hoover-bot"
version = "0.1.0"
description = "Listen To This bot"
authors = ["codyconfer <mail@codyconfer.me>"]
readme = "README.md"
packages = [{include = "./src"}]

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.2.2"
python-dotenv = "^1.0.0"
pydantic = "^1.10.7"
black = "^23.3.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
